backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
aiofiles
requests
pydantic
python-multipart
openai
google-generativeai
gTTS
pydub
bcrypt
pyjwt
pytest
pytest-asyncio
httpx
